+++ scripts/macros.js
@@ -1,4 +1,5 @@
 export {acidSplash} from './macros/2024/spells/acidSplash.js';
 export {aid} from './macros/2024/spells/aid.js';
 export {alterSelf} from './macros/2024/spells/alterSelf.js';
+export {animalFriendship} from './macros/2024/spells/animalFriendship.js';
 export {armorOfAgathys} from './macros/2024/spells/armorOfAgathys.js';